test deserialize
